change the name and the concept of sending emails ,add activation email sender
