--- backend/utils/email.js
@@ -1,29 +1,0 @@
-import nodemailer from 'nodemailer';
-
-const sendEmail = async (email,resetLink) => {
-  const transporter = nodemailer.createTransport({
-    service: 'gmail',
-    auth: {
-      user: process.env.EMAIL,
-      pass: process.env.EMAIL_PASSWORD,
-    },
-  });
-
-  const mailOptions = {
-    from: process.env.EMAIL,
-    to: email,
-    subject: "Password Reset",
-    text: `Click the following link to reset your password: ${resetLink}`,
-    html: `<p>Click the following link to reset your password: <a href="${resetLink}">${resetLink}</a></p>`,
-  };
-
-  try {
-    await transporter.sendMail(mailOptions);
-    console.log('Password reset email sent');
-  } catch (error) {
-    console.error('Error sending password reset email:', error);
-    throw new Error('Error sending password reset email');
-  }
-};
-
-export default sendEmail;
